refactor: fix review issues — mark mocked responses, consolidate VM Service funcs, cleanup dead code, bump v0.3.0

+++ src/ui/network/detail.rs
@@ -111,7 +111,7 @@ pub fn draw_network_detail(f: &mut Frame, app: &mut App, area: Rect) {
             // Register click region — will be offset by area.x+1 and the content_y + line_idx
             // We store the raw line index; the click handler translates using detail_content_y + scroll
             let line_idx = all_lines.len();
-            let btn_x_start = (area.x + 1 + divider_w as u16) as u16;
+            let btn_x_start = area.x + 1 + divider_w as u16;
             let btn_x_end = btn_x_start + mock_btn.len() as u16;
             let btn_y = area.y + 1 + line_idx as u16; // approximate — before scroll
             app.layout.detail_mock_btn = Some((btn_y, btn_x_start, btn_x_end));
